Flere kommentarer og lidt oprydning

+++ src/privatemoviecollection/GUI/Controller/FXMLDocumentController.java
@@ -90,19 +90,18 @@ public class FXMLDocumentController implements Initializable
         searchDone = false;
     }
 
+    /*
+    Creates a new object of the MovieModel class and populates the tableviews
+    */
     @Override
     public void initialize(URL url, ResourceBundle rb)
     {
-
         try
         {
             mModel = new MovieModel();
             tbViewMovie.setItems(mModel.getAllMovies());
             tbViewCategory.setItems(mModel.getAllCategories());
-        } catch (IOException ex)
-        {
-            Logger.getLogger(FXMLDocumentController.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (SQLException ex)
+        } catch (IOException | SQLException ex)
         {
             Logger.getLogger(FXMLDocumentController.class.getName()).log(Level.SEVERE, null, ex);
         }
@@ -117,6 +116,10 @@ public class FXMLDocumentController implements Initializable
 
     }
 
+    /*
+    sends a search input to the model and inserts the responce into the tableview
+    the second click will clear the input field.
+    */
     @FXML
     public void handleSearch(ActionEvent event) throws SQLException
     {
@@ -134,7 +137,9 @@ public class FXMLDocumentController implements Initializable
             txtSearch.clear();
         }
     }
-
+    /*
+    creates a new category if the input is valid
+    */
     @FXML
     public void addCategory(ActionEvent event)
     {
@@ -147,7 +152,10 @@ public class FXMLDocumentController implements Initializable
         }
         mModel.addCategory(newCat);
     }
-
+    /*
+    opens a new window which is then used to add a new movie
+    @throws IOException
+    */
     @FXML
     public void addMovie(ActionEvent event) throws IOException
     {
@@ -165,18 +173,28 @@ public class FXMLDocumentController implements Initializable
         stageAddMovie.initOwner(primeStage);
         stageAddMovie.show();
     }
-
+    
+    /*
+    Deletes the selected movie
+    */
     @FXML
     public void deleteMovie() throws SQLException
     {
+        if (tbViewMovie.getSelectionModel().isEmpty())
+        {
+            JOptionPane.showMessageDialog(null, "You have to select a movie to delete", "Invalid selection", JOptionPane.ERROR_MESSAGE);
+            return;
+        }
         int p = JOptionPane.showConfirmDialog(null, "Do you want to delete this movie?", "Delete", JOptionPane.YES_NO_OPTION);
         if (p == 0)
         {
             mModel.deleteMovie(tbViewMovie.getSelectionModel().getSelectedItem());
-
         }
     }
-
+    
+    /*
+    Deletes the selected category
+    */
     @FXML
     public void deleteCategory(ActionEvent event)
     {
@@ -192,10 +210,17 @@ public class FXMLDocumentController implements Initializable
             mModel.deleteCategory(tbViewCategory.getSelectionModel().getSelectedItem());
         }
     }
-
+    /*
+    adds a personal rating to the selected movie between 1 and 10
+    */
     @FXML
     public void addPersonalRating(ActionEvent event)
     {
+        if (tbViewMovie.getSelectionModel().isEmpty())
+        {
+            JOptionPane.showMessageDialog(null, "you have to select a movie", "invalid selection", JOptionPane.ERROR_MESSAGE);
+           return;
+        }
         double p = Double.parseDouble(JOptionPane.showInputDialog(null, "Enter a rating between 1 - 10", "Enter", JOptionPane.OK_CANCEL_OPTION));
 
         if (p >= 1 && p <= 10)
@@ -206,7 +231,10 @@ public class FXMLDocumentController implements Initializable
             JOptionPane.showMessageDialog(null, "You have to enter a number between 1 and 10", "Incorrect number", JOptionPane.ERROR_MESSAGE);
         }
     }
-
+    
+    /*
+    plays the tableview movie when you click it twice
+    */
     @FXML
     public void clickPlayMovie(MouseEvent event) throws IOException
     {
@@ -228,18 +256,10 @@ public class FXMLDocumentController implements Initializable
             }
         }
     }
-public Movie selectTableviewMovies(Movie movieToWatch)
-{
-    if (!tbViewMovie.getSelectionModel().isEmpty())
-    {
-        movieToWatch = tbViewMovie.getSelectionModel().getSelectedItem();
-    }
-    else if (!tbViewCatMovies.getSelectionModel().isEmpty())
-    {
-        movieToWatch = tbViewCatMovies.getSelectionModel().getSelectedItem();
-    }
-    return movieToWatch;
-}
+    
+    /*
+    opens imdb's homepage
+    */
     @FXML
     public void openImdb(ActionEvent event)
     {
@@ -247,12 +267,15 @@ public Movie selectTableviewMovies(Movie movieToWatch)
         {
             String url = "https://www.imdb.com/";
             Desktop.getDesktop().browse(URI.create(url));
-        } catch (Exception e)
+        } catch (IOException e)
         {
             e.printStackTrace();
         }
     }
 
+    /*
+    selects which category's items are shown
+    */
     @FXML
     public void selectCat(MouseEvent event) throws SQLException
     {
@@ -271,7 +294,10 @@ public Movie selectTableviewMovies(Movie movieToWatch)
             }
         }
     }
-
+    
+    /*
+    adds selected movie to the selected category
+    */
     @FXML
     public void addToCategory(ActionEvent event)
     {
@@ -285,9 +311,16 @@ public Movie selectTableviewMovies(Movie movieToWatch)
         mModel.addToCategory(selectedMovie, selectedCat);
     }
 
+    /*
+    removes the selected movie from the selected category
+    */
     @FXML
     public void deleteFromCategory(ActionEvent event)
     {
+        if (tbViewCatMovies.getSelectionModel().getSelectedItem() == null) {
+            JOptionPane.showMessageDialog(null, "You have to select a movie from the list of category movies to delete", "Illegal selection", JOptionPane.ERROR_MESSAGE);
+            return;
+        }
         int p = JOptionPane.showConfirmDialog(null, "Do you want to delete this movie from this category", "Delete", JOptionPane.YES_NO_OPTION);
 
         if (p == 0)
@@ -297,10 +330,4 @@ public Movie selectTableviewMovies(Movie movieToWatch)
             tbViewCatMovies.setItems(mModel.getAllMoviesInCategory(currentCategory)); // gets the new list of songs minus the deleted ones
         }
     }
-
-    public void refreshTableview()
-    {
-        tbViewMovie.refresh();
-    }
-
 }
